// src/interfaces/Seller.interface.ts
export interface ISeller {
  id?: string;
  seller_id?: string;
  email_id?: string;
  storeName?: string;
  sellerName?: string;
  status?: string;
  city?: string;
  product_category?: string;
  storeStatus?: string;
  storeCreatedAt?: string;
  reviewComment?: string;
  phoneNumber?: string;
}